move md5 out to index
