+++ mir_dataset_loaders/download_utils.py
@@ -1,5 +1,4 @@
 from collections import namedtuple
-import hashlib
 import os
 import tarfile
 from urllib import request
@@ -8,6 +7,7 @@ import zipfile
 from tqdm import tqdm
 
 from . import MIR_DATASETS_DIR
+from . import md5
 
 RemoteFileMetadata = namedtuple('RemoteFileMetadata',
                                 ['filename', 'url', 'checksum'])
@@ -37,26 +37,6 @@ def get_save_path(data_home):
         os.makedirs(save_path)
 
     return save_path
-
-
-def md5(file_path):
-    """Get md5 hash of a file.
-
-    Parameters
-    ----------
-    file_path: str
-        File path.
-
-    Returns
-    -------
-    md5_hash: str
-        md5 hash of data in file_path
-    """
-    hash_md5 = hashlib.md5()
-    with open(file_path, "rb") as fhandle:
-        for chunk in iter(lambda: fhandle.read(4096), b""):
-            hash_md5.update(chunk)
-    return hash_md5.hexdigest()
 
 
 class DownloadProgressBar(tqdm):
